pyproject.toml
[project]
name = "ez-image"
version = "0.1.0"
description = "Image generation and editing skills for AI agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-auth>=2.49.1",
    "google-genai>=1.68.0",
    "httpx>=0.27.0",
    "pillow>=12.1.1",
    "typer>=0.24.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.15.4",
    "prek>=0.3.4",
    "pytest",
    "pytest-asyncio>=1.3.0",
    "pytest-cov",
    "ty>=0.0.19",
]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
]
markers = [
    "e2e: end-to-end tests that call real APIs (slow, requires credentials)",
]

[tool.coverage.run]
source = ["skills/ez-google-imagen/scripts"]
omit = ["*/tests/*"]

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "if __name__ == .__main__.",
]

[tool.isort]
profile = "black"
src_paths = "tests"
line_length = 80
include_trailing_comma = true
multi_line_output = 3

[tool.ruff]
select = ["E", "F", "B"]
exclude = ["__pycache__", "terraform", ".git", ".venv"]
ignore = ["E501"]
